+++ examples/asyncio/sanitize/sanitize_examples/sanitize_no_share.py
@@ -9,7 +9,7 @@ from pangea.response import TransferMethod
 from pangea.services.sanitize import SanitizeContent, SanitizeFile, SanitizeShareOutput
 
 # Set this filepath to your own file
-FILEPATH = "./sanitize_examples/ds11.pdf"
+FILEPATH = "./sanitize_examples/test-sanitize.txt"
 
 
 async def main() -> None:
@@ -35,8 +35,6 @@ async def main() -> None:
             domain_intel_provider="crowdstrike",
             defang=True,
             defang_threshold=20,
-            remove_interactive=True,
-            remove_attachments=True,
             redact=True,
         )
         # Disable share output
@@ -60,7 +58,6 @@ async def main() -> None:
 
             print(f"\tRedact data: {response.result.data.redact}")
             print(f"\tDefang data: {response.result.data.defang}")
-            print(f"\tCDR data: {response.result.data.cdr}")
 
             if response.result.data.malicious_file:
                 print("File IS malicious")