+++ packages/pangea-sdk/tests/integration/asyncio/test_sanitize.py
@@ -5,6 +5,8 @@ import unittest
 from asyncio import sleep
 from contextlib import suppress
 from http.client import HTTPConnection
+from io import BufferedReader
+from pathlib import Path
 
 import pangea.exceptions as pe
 from pangea import PangeaConfig
@@ -17,14 +19,14 @@ from pangea.utils import get_file_upload_params
 from tests.test_tools import load_test_environment
 
 TEST_ENVIRONMENT = load_test_environment(SanitizeAsync.service_name, TestEnvironment.LIVE)
-PDF_FILEPATH = "./tests/testdata/ds11.pdf"
+PDF_FILEPATH = Path("./tests/testdata/test-sanitize.txt")
 
 
-def get_test_file():
-    return open(PDF_FILEPATH, "rb")
+def get_test_file() -> BufferedReader:
+    return PDF_FILEPATH.open("rb")
 
 
-def debug_requests_on():
+def debug_requests_on() -> None:
     """Switches on logging of the requests module."""
     HTTPConnection.debuglevel = 1
 
@@ -49,10 +51,10 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
         self.client = SanitizeAsync(token, config=config)
         logger_set_pangea_config(logger_name=self.client.logger.name)
 
-    async def asyncTearDown(self):
+    async def asyncTearDown(self) -> None:
         await self.client.close()
 
-    async def test_sanitize_and_share(self):
+    async def test_sanitize_and_share(self) -> None:
         file_scan = SanitizeFile(scan_provider="crowdstrike")
         content = SanitizeContent(
             url_intel=True,
@@ -61,8 +63,6 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
             domain_intel_provider="crowdstrike",
             defang=True,
             defang_threshold=20,
-            remove_interactive=True,
-            remove_attachments=True,
             redact=True,
         )
         share_output = SanitizeShareOutput(enabled=True, output_folder="sdk_test/sanitize/")
@@ -76,19 +76,23 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
                 uploaded_file_name="uploaded_file",
             )
             self.assertEqual(response.status, "Success")
+            assert response.result
             self.assertIsNone(response.result.dest_url)
             self.assertIsNotNone(response.result.dest_share_id)
+            assert response.result.data.redact
             self.assertGreater(response.result.data.redact.redaction_count, 0)
             self.assertNotEqual(response.result.data.redact.summary_counts, {})
+            assert response.result.data.defang
+            assert response.result.data.defang.external_urls_count
             self.assertGreater(response.result.data.defang.external_urls_count, 0)
+            assert response.result.data.defang.external_domains_count
             self.assertGreater(response.result.data.defang.external_domains_count, 0)
+            assert response.result.data.defang.defanged_count is not None
             self.assertEqual(response.result.data.defang.defanged_count, 0)
             self.assertIsNotNone(response.result.data.defang.domain_intel_summary)
-            self.assertEqual(response.result.data.cdr.file_attachments_removed, 0)
-            self.assertEqual(response.result.data.cdr.interactive_contents_removed, 0)
             self.assertFalse(response.result.data.malicious_file)
 
-    async def test_sanitize_no_share(self):
+    async def test_sanitize_no_share(self) -> None:
         file_scan = SanitizeFile(scan_provider="crowdstrike")
         content = SanitizeContent(
             url_intel=True,
@@ -97,8 +101,6 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
             domain_intel_provider="crowdstrike",
             defang=True,
             defang_threshold=20,
-            remove_interactive=True,
-            remove_attachments=True,
             redact=True,
         )
         share_output = SanitizeShareOutput(enabled=False)
@@ -112,22 +114,25 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
                 uploaded_file_name="uploaded_file",
             )
             self.assertEqual(response.status, "Success")
-            self.assertIsNotNone(response.result.dest_url)
+            assert response.result
+            assert response.result.dest_url
             self.assertIsNone(response.result.dest_share_id)
+            assert response.result.data.redact
             self.assertGreater(response.result.data.redact.redaction_count, 0)
             self.assertNotEqual(response.result.data.redact.summary_counts, {})
+            assert response.result.data.defang
+            assert response.result.data.defang.external_urls_count
             self.assertGreater(response.result.data.defang.external_urls_count, 0)
+            assert response.result.data.defang.external_domains_count
             self.assertGreater(response.result.data.defang.external_domains_count, 0)
             self.assertEqual(response.result.data.defang.defanged_count, 0)
             self.assertIsNotNone(response.result.data.defang.domain_intel_summary)
-            self.assertEqual(response.result.data.cdr.file_attachments_removed, 0)
-            self.assertEqual(response.result.data.cdr.interactive_contents_removed, 0)
             self.assertFalse(response.result.data.malicious_file)
 
             attached_file = await self.client.download_file(response.result.dest_url)
             attached_file.save("./")
 
-    async def test_sanitize_all_defaults(self):
+    async def test_sanitize_all_defaults(self) -> None:
         with suppress(pe.AcceptedRequestException), get_test_file() as f:
             response = await self.client.sanitize(
                 file=f,
@@ -135,15 +140,14 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
                 uploaded_file_name="uploaded_file",
             )
             self.assertEqual(response.status, "Success")
+            assert response.result
             self.assertIsNotNone(response.result.dest_url)
             self.assertIsNone(response.result.dest_share_id)
             self.assertIsNone(response.result.data.redact)
             self.assertIsNotNone(response.result.data.defang)
-            self.assertEqual(response.result.data.cdr.file_attachments_removed, 0)
-            self.assertEqual(response.result.data.cdr.interactive_contents_removed, 0)
             self.assertFalse(response.result.data.malicious_file)
 
-    async def test_sanitize_multipart_upload(self):
+    async def test_sanitize_multipart_upload(self) -> None:
         file_scan = SanitizeFile(scan_provider="crowdstrike")
         content = SanitizeContent(
             url_intel=True,
@@ -152,8 +156,6 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
             domain_intel_provider="crowdstrike",
             defang=True,
             defang_threshold=20,
-            remove_interactive=True,
-            remove_attachments=True,
             redact=True,
         )
         share_output = SanitizeShareOutput(enabled=True, output_folder="sdk_test/sanitize/")
@@ -167,26 +169,28 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
                 uploaded_file_name="uploaded_file",
             )
             self.assertEqual(response.status, "Success")
+            assert response.result
             self.assertIsNone(response.result.dest_url)
             self.assertIsNotNone(response.result.dest_share_id)
+            assert response.result.data.redact
             self.assertGreater(response.result.data.redact.redaction_count, 0)
             self.assertNotEqual(response.result.data.redact.summary_counts, {})
+            assert response.result.data.defang
+            assert response.result.data.defang.external_urls_count
             self.assertGreater(response.result.data.defang.external_urls_count, 0)
+            assert response.result.data.defang.external_domains_count
             self.assertGreater(response.result.data.defang.external_domains_count, 0)
             self.assertEqual(response.result.data.defang.defanged_count, 0)
             self.assertIsNotNone(response.result.data.defang.domain_intel_summary)
-            self.assertEqual(response.result.data.cdr.file_attachments_removed, 0)
-            self.assertEqual(response.result.data.cdr.interactive_contents_removed, 0)
             self.assertFalse(response.result.data.malicious_file)
 
-    async def test_sanitize_async(self):
-        with self.assertRaises(pe.AcceptedRequestException):
-            with get_test_file() as f:
-                response = await self.client.sanitize(
-                    file=f, transfer_method=TransferMethod.POST_URL, uploaded_file_name="uploaded_file", sync_call=False
-                )
+    async def test_sanitize_async(self) -> None:
+        with self.assertRaises(pe.AcceptedRequestException), get_test_file() as f:
+            await self.client.sanitize(
+                file=f, transfer_method=TransferMethod.POST_URL, uploaded_file_name="uploaded_file", sync_call=False
+            )
 
-    async def test_sanitize_filepath(self):
+    async def test_sanitize_filepath(self) -> None:
         with suppress(pe.AcceptedRequestException):
             file_scan = SanitizeFile(scan_provider="crowdstrike")
             content = SanitizeContent(
@@ -196,13 +200,11 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
                 domain_intel_provider="crowdstrike",
                 defang=True,
                 defang_threshold=20,
-                remove_interactive=True,
-                remove_attachments=True,
                 redact=True,
             )
             share_output = SanitizeShareOutput(enabled=True, output_folder="sdk_test/sanitize/")
             response = await self.client.sanitize(
-                file_path=PDF_FILEPATH,
+                file_path=str(PDF_FILEPATH),
                 transfer_method=TransferMethod.MULTIPART,
                 file_scan=file_scan,
                 content=content,
@@ -210,21 +212,22 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
                 uploaded_file_name="uploaded_file",
             )
             self.assertEqual(response.status, "Success")
+            assert response.result
             self.assertIsNone(response.result.dest_url)
             self.assertIsNotNone(response.result.dest_share_id)
-            self.assertIsNotNone(response.result.data.redact)
+            assert response.result.data.redact
             self.assertGreater(response.result.data.redact.redaction_count, 0)
             self.assertNotEqual(response.result.data.redact.summary_counts, {})
-            self.assertIsNotNone(response.result.data.defang)
+            assert response.result.data.defang
+            assert response.result.data.defang.external_urls_count
             self.assertGreater(response.result.data.defang.external_urls_count, 0)
+            assert response.result.data.defang.external_domains_count
             self.assertGreater(response.result.data.defang.external_domains_count, 0)
             self.assertEqual(response.result.data.defang.defanged_count, 0)
             self.assertIsNotNone(response.result.data.defang.domain_intel_summary)
-            self.assertEqual(response.result.data.cdr.file_attachments_removed, 0)
-            self.assertEqual(response.result.data.cdr.interactive_contents_removed, 0)
             self.assertFalse(response.result.data.malicious_file)
 
-    async def test_sanitize_poll_result(self):
+    async def test_sanitize_poll_result(self) -> None:
         exception = None
         try:
             with get_test_file() as f:
@@ -243,22 +246,21 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
                 # wait some time to get result ready and poll it
                 await sleep(10)
 
-                response: PangeaResponse[SanitizeResult] = await self.client.poll_result(exception)
+                response: PangeaResponse[SanitizeResult] = await self.client.poll_result(exception)  # type: ignore[no-redef]
                 self.assertEqual(response.status, "Success")
+                assert response.result
                 self.assertIsNotNone(response.result.dest_url)
                 self.assertIsNone(response.result.dest_share_id)
                 self.assertIsNone(response.result.data.redact)
                 self.assertIsNotNone(response.result.data.defang)
-                self.assertEqual(response.result.data.cdr.file_attachments_removed, 0)
-                self.assertEqual(response.result.data.cdr.interactive_contents_removed, 0)
                 self.assertFalse(response.result.data.malicious_file)
                 return
             except pe.AcceptedRequestException:
                 pass
 
-        self.log.warning("The result of request '%s' took too long to be ready.", exception.request_id)
+        self.log.warning("The result of request '%s' took too long to be ready.", exception.request_id)  # type: ignore[union-attr]
 
-    async def test_split_upload_file_post(self):
+    async def test_split_upload_file_post(self) -> None:
         file_scan = SanitizeFile(scan_provider="crowdstrike")
         content = SanitizeContent(
             url_intel=True,
@@ -267,8 +269,6 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
             domain_intel_provider="crowdstrike",
             defang=True,
             defang_threshold=20,
-            remove_interactive=True,
-            remove_attachments=True,
             redact=True,
         )
         share_output = SanitizeShareOutput(enabled=False)
@@ -282,6 +282,8 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
                 params=params,
                 uploaded_file_name="uploaded_file",
             )
+            assert response.accepted_result
+            assert response.accepted_result.post_url
             url = response.accepted_result.post_url
             file_details = response.accepted_result.post_form_data
 
@@ -296,21 +298,21 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
                 # wait some time to get result ready and poll it
                 await sleep(10)
 
-                response: PangeaResponse[SanitizeResult] = await self.client.poll_result(response=response)
+                response: PangeaResponse[SanitizeResult] = await self.client.poll_result(response=response)  # type: ignore[no-redef]
                 self.assertEqual(response.status, "Success")
+                assert response.result
                 self.assertIsNotNone(response.result.dest_url)
                 self.assertIsNone(response.result.dest_share_id)
-                self.assertIsNotNone(response.result.data.redact)
+                assert response.result.data.redact
                 self.assertGreater(response.result.data.redact.redaction_count, 0)
                 self.assertNotEqual(response.result.data.redact.summary_counts, {})
-                self.assertIsNotNone(response.result.data.defang)
+                assert response.result.data.defang
+                assert response.result.data.defang.external_urls_count
                 self.assertGreater(response.result.data.defang.external_urls_count, 0)
+                assert response.result.data.defang.external_domains_count
                 self.assertGreater(response.result.data.defang.external_domains_count, 0)
                 self.assertEqual(response.result.data.defang.defanged_count, 0)
                 self.assertIsNotNone(response.result.data.defang.domain_intel_summary)
-                self.assertEqual(response.result.data.cdr.file_attachments_removed, 0)
-                self.assertEqual(response.result.data.cdr.file_attachments_removed, 0)
-                self.assertEqual(response.result.data.cdr.interactive_contents_removed, 0)
                 self.assertFalse(response.result.data.malicious_file)
                 return
             except pe.AcceptedRequestException:
@@ -318,7 +320,7 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
 
         self.log.warning("The result of request '%s' took too long to be ready.", response.request_id)
 
-    async def test_split_upload_file_put(self):
+    async def test_split_upload_file_put(self) -> None:
         file_scan = SanitizeFile(scan_provider="crowdstrike")
         content = SanitizeContent(
             url_intel=True,
@@ -327,8 +329,6 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
             domain_intel_provider="crowdstrike",
             defang=True,
             defang_threshold=20,
-            remove_interactive=True,
-            remove_attachments=True,
             redact=True,
         )
         share_output = SanitizeShareOutput(enabled=False)
@@ -340,6 +340,8 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
                 share_output=share_output,
                 uploaded_file_name="uploaded_file",
             )
+            assert response.accepted_result
+            assert response.accepted_result.put_url
             url = response.accepted_result.put_url
 
             uploader = FileUploaderAsync()
@@ -351,20 +353,21 @@ class TestSanitize(unittest.IsolatedAsyncioTestCase):
                 # wait some time to get result ready and poll it
                 await sleep(10)
 
-                response: PangeaResponse[SanitizeResult] = await self.client.poll_result(response=response)
+                response: PangeaResponse[SanitizeResult] = await self.client.poll_result(response=response)  # type: ignore[no-redef]
                 self.assertEqual(response.status, "Success")
+                assert response.result
                 self.assertIsNotNone(response.result.dest_url)
                 self.assertIsNone(response.result.dest_share_id)
-                self.assertIsNotNone(response.result.data.redact)
+                assert response.result.data.redact
                 self.assertGreater(response.result.data.redact.redaction_count, 0)
                 self.assertNotEqual(response.result.data.redact.summary_counts, {})
-                self.assertIsNotNone(response.result.data.defang)
+                assert response.result.data.defang
+                assert response.result.data.defang.external_urls_count
                 self.assertGreater(response.result.data.defang.external_urls_count, 0)
+                assert response.result.data.defang.external_domains_count
                 self.assertGreater(response.result.data.defang.external_domains_count, 0)
                 self.assertEqual(response.result.data.defang.defanged_count, 0)
                 self.assertIsNotNone(response.result.data.defang.domain_intel_summary)
-                self.assertEqual(response.result.data.cdr.file_attachments_removed, 0)
-                self.assertEqual(response.result.data.cdr.interactive_contents_removed, 0)
                 self.assertFalse(response.result.data.malicious_file)
                 return
             except pe.AcceptedRequestException: