+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Removed
+
+- CDR and PDF support in Sanitize.
+
 ## 5.3.0 - 2025-01-13
 
 ### Added