+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - AI Guard API response models.
 
+### Removed
+
+- CDR and PDF support in Sanitize.
+
 ## 5.3.0 - 2025-01-13
 
 ### Added