+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- "info" and "classifications" fields in Prompt Guard.
+
 ### Changed
 
 - Renamed "detector" to "analyzer" in Prompt Guard.