+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 5.4.0beta2 - 2025-01-21
+
 ### Added
 
 - "info" and "classifications" fields in Prompt Guard.